+++ src/dream/QueryDriver/helpers/kysely/runMigration.ts
@@ -1,6 +1,7 @@
 import { FileMigrationProvider, MigrationResult, Migrator } from 'kysely'
 import * as fs from 'node:fs/promises'
 import * as path from 'node:path'
+import DreamCLI from '../../../../cli/index.js'
 import colorize from '../../../../cli/logger/loggable/colorize.js'
 import {
   closeAllConnectionsForConnectionName,
@@ -9,7 +10,7 @@ import {
 } from '../../../../db/DreamDbConnection.js'
 import db from '../../../../db/index.js'
 import DreamApp from '../../../../dream-app/index.js'
-import DreamCLI from '../../../../cli/index.js'
+import migrationFolderPath from './migrationFolderPath.js'
 
 type MigrationModes = 'migrate' | 'rollback'
 
@@ -24,11 +25,7 @@ export default async function runMigration({
   mode = 'migrate',
   dialectProvider,
 }: MigrationOpts) {
-  const dreamApp = DreamApp.getOrFail()
-  const migrationFolder =
-    connectionName === 'default'
-      ? path.join(dreamApp.projectRoot, dreamApp.paths.db, 'migrations')
-      : path.join(dreamApp.projectRoot, dreamApp.paths.db, 'migrations', connectionName)
+  const migrationFolder = migrationFolderPath(connectionName)
 
   // Ensure the migration folder exists
   await fs.mkdir(migrationFolder, { recursive: true })